Se creo la pagina estatica our-values que contiene la mision, vision y valores, se modifico la pagina "Sobre nosotros", para que ya no contenga esta seccion.
